updated readme; cleanup linop; renamed solver.py to linalg.py and added parallel_norm to it from tools.py
