+++ brahmap/math/linalg.py
@@ -2,13 +2,21 @@ import numpy as np
 import scipy
 import scipy.sparse
 import scipy.sparse.linalg
-import brahmap
+
+from brahmap import MPI_UTILS
+
+
+def parallel_norm(x: np.ndarray):
+    sqnorm = x.dot(x)
+    sqnorm = MPI_UTILS.comm.allreduce(sqnorm)
+    ret = np.sqrt(sqnorm)
+    return ret
 
 
 def cg(A, b, x0=None, atol=1.0e-12, maxiter=100, M=None, callback=None):
     A, M, x, b, postprocess = scipy.sparse.linalg._isolve.utils.make_system(A, M, x0, b)
 
-    b_norm = brahmap.utilities.parallel_norm(b)
+    b_norm = parallel_norm(b)
 
     if b_norm == 0:
         return postprocess(b), 0
@@ -43,7 +51,7 @@ def cg(A, b, x0=None, atol=1.0e-12, maxiter=100, M=None, callback=None):
         r -= alpha * q
         rho_prev = rho_cur
 
-        norm_residual = brahmap.utilities.parallel_norm(r) / b_norm
+        norm_residual = parallel_norm(r) / b_norm
 
         if callback:
             callback(x, r, norm_residual)
