+++ docs/api_reference/math_functions/parallel_norm.md
@@ -1,0 +1,3 @@
+# `parallel_norm`
+
+::: brahmap.math.parallel_norm
