+++ docs/api_reference/utilities/index.md
@@ -6,7 +6,6 @@
 
 ## Tools
 
-- [`parallel_norm`](parallel_norm.md)
 - [`modify_numpy_context`](modify_numpy_context.md)
 - [`TypeChangeWarning`](TypeChangeWarning.md)
 - [`LowerTypeCastWarning`](LowerTypeCastWarning.md)
