--- docs/api_reference/utilities/parallel_norm.md
@@ -1,3 +1,0 @@
-# `parallel_norm`
-
-::: brahmap.utilities.parallel_norm
